Add files via upload
changeADAdminGroups.sh -- Change the AD Administrative Groups

This script will modify the groups from an Active Directory domain controller that will have administrative access on the machine. This script should be run after a machine has been bound to Active Directory. The <timeout> value can be used with a hardcoded value in the script, or read in as a parameter. Since the Casper Suite defines the first three parameters as (1) Mount Point, (2) Computer Name and (3) username, we are using the fourth parameter ($4) as the passable parameter.